requirement.txt
python-dotenv>=1.0.0
openai>=1.0.0
mcp>=0.1.0
rich>=13.0.0
